added status page for cleanup
