+++ pages/admin/signUp/error.tsx
@@ -1,42 +1,3 @@
-import { Grid, Spacer, Input, Button, Text} from "@nextui-org/react";
-import { useRouter } from "next/router";
-import styles from "../admin.module.css";
+import StatusPage from "../../../components/admin/dashboard/statusPage";
 
-
-const SignupError = () => {
-    
-    const router = useRouter();
-    
-    return (
-        <div className={styles.page}>
-            <div className={styles.successContainer}>
-               
-
-                    <Text className={styles.successText}>The invite link used is either invalid or it expired. Please ask the administrators for a new invite.</Text>
-
-                    <Spacer y={0.75} />
-
-                   
-
-                    <Spacer y={0.75} />
-
-                    <Button
-                        id="login"
-                        className={styles.loginButton}
-                        type="submit"
-                        onPress={() => {router.push("/admin");}}
-                    >
-                        <Text className={styles.loginButtonText}>Go to login</Text>
-                    </Button>
-
-                    <Spacer y={0.5} />
-
-                   
-
-               
-            </div>
-        </div>
-    )
-}
-
-export default SignupError;
+export default <StatusPage message="The invite link used is either invalid or it expired. Please ask the administrators for a new invite."/>
