+++ pages/admin/signUp/index.tsx
@@ -1,5 +1,3 @@
 import SignupError from "./error";
 
-export default function SignUpIndex() {
-    return <SignupError/>
-}
+export default SignupError
